Fix unzoom not working properly in Edge
There seems to be a problem with nested transitions in Edge - when the
wrapper translates, the image's scale transition gets cancelled. Due to
this, the `transitionend` event doesn't get fired, causing `dispose` to
not be called.

This fixes the issue by attaching events to the wrapper's
`transitionend`, which does get called.

CodePen: https://codepen.io/spinningarrow/pen/EEJwzN
SO: https://stackoverflow.com/q/49799116/342886